alpha release

+++ yemek_koala-master/db/migrate/20160810185552_change_id_from_event.rb
@@ -1,5 +1,5 @@
 class ChangeIdFromEvent < ActiveRecord::Migration
   def change
-    change_column :events, :id, :integer, limit: 8
+    change_column :events, :id, :bigint, limit: 8
   end
 end
